master ブランチ 更新サマリ (2026-06-27 〜 2026-07-19)
■ Emscripten (wasm32) ビルド対応
  - Emscripten (wasm32) ターゲットのビルドに対応 (CMake 対応 + Bin2C
    リソース変換 + SDL3 メインループの wasm 適合)
  - web:// / webnc:// ストレージメディアを追加。バラファイルを fetch で
    取得し、OPFS 永続キャッシュに対応。manifest ベースの GetListAt /
    存在確認、ファイル名列挙の UTF-8→UTF-16 変換、プロジェクトパスへの
    直接指定に対応
  - リソースを wasm バイナリ非埋め込みにし、外部 MEMFS /resource から
    読み込むよう変更
  - devtools コンソールから使える REPL ブリッジを追加
  - Elements ダイアログを wasm で利用可能に (overlay 系 + JSPI による
    モーダル対応)

■ REPL 拡張
  - abstract unix socket チャネルを追加 (Android/Linux)。外部プロセス
    からの REPL 接続が可能に
  - socket チャネルの dot コマンド (.help/.depth/.compact) と複数行継続
    に対応

■ 動画 / サウンド
  - Android で動画音声が遅延する問題を修正 (MovieAudioSinkAdapter:
    codec buffer のコピー + 流量制御)。movie-player submodule も更新
  - 存在しないファイルのオープンが TVPUnknownWaveFormat と誤報される
    問題を修正 (正しく「ファイルが見つからない」を報告)

■ エラーハンドリング
  - 未処理スクリプト例外の後処理をプラットフォーム別ポリシー化
    (従来の即終了に加え、継続実行などプラットフォームに応じた挙動を選択可能に)

■ ビルド / リソース
  - リソースフォルダを KRKRZ_RESOURCE_DIR で差し替え可能に
  - x86 (32bit) ビルド修正 (呼び出し規約の不整合 + AllocStats ヘッダサイズ)
  - .gitattributes 調整 (文字コード指定カット / .rc の指定除外)

■ 機能追加 / 修正
  - opengl: Offscreen FBO にステンシルバッファを追加
  - plugin: 静的プラグインの link/unlink で登録名をキー正規化
  - pad: 全パッドを開いたまま保持し、最後に押したパッドへメインを切替
    できるように
  - tp_stub: TVP_WM_USER を全バリアントで定義しユーザメッセージ基点を統一
  - elements: Android でダイアログの IME を focus 駆動に (submodule 更新)

■ 過去リリースからの削除追従 (反映漏れ整理)
  - 上流で削除済みだったファイルを本ミラーからも削除:
    WaveFormatConverter / XAudio2AudioStream / sdl3 旧 AudioStream /
    NativeEventQueue / ApplicationSpecialPath / FilePathUtil /
    IA32 NASM ソース (tlg6_*.nas, nasm.nah)

+++ common/utils/GlobalAllocStats.cpp
@@ -62,12 +62,12 @@ constexpr uint32_t kMagicRaw  = 0x4D454D31u; // "1MEM"  (32-bit short)
 constexpr uint32_t kMagicPool = 0x4D454D32u; // "2MEM"
 
 struct Header {
-	size_t   size;     // 8
+	uint64_t size;     // 8 (size_t だと 32bit 環境で 4 byte になり 16 byte 前提が崩れる)
 	uint32_t magic;    // 4
 	uint16_t tag;      // 2 (TVPAllocTag as raw uint16)
 	uint16_t pad;      // 2 (alignment, unused)
 };
-// MSVC x64 / Linux x86_64 共通で size_t = 8 byte なので Header = 16 byte。
+// size を明示 64bit にしているため 32bit / 64bit いずれの環境でも Header = 16 byte。
 // __STDCPP_DEFAULT_NEW_ALIGNMENT__ (典型 8 or 16) を満たすため 16 倍数を確認。
 static_assert(sizeof(Header) == 16, "Header must be exactly 16 bytes");
 static_assert(sizeof(Header) % 16 == 0, "Header must be 16-byte aligned");
@@ -265,7 +265,7 @@ inline void do_free(Collector &c, void *p) {
 	// post-init モード: header magic で経路を識別。
 	Header *h = reinterpret_cast<Header *>(static_cast<char *>(p) - sizeof(Header));
 	if (h->magic == kMagicPool) {
-		size_t size = h->size;
+		size_t size = static_cast<size_t>(h->size);
 		uint16_t tag = h->tag;
 		c.recordFree(size, tag);
 		// pool は bindPool 後 unbind しない前提。
@@ -278,7 +278,7 @@ inline void do_free(Collector &c, void *p) {
 		return;
 	}
 	if (h->magic == kMagicRaw) {
-		c.recordFree(h->size, h->tag);
+		c.recordFree(static_cast<size_t>(h->size), h->tag);
 		std::free(h);
 		return;
 	}
@@ -313,7 +313,7 @@ inline void *do_realloc(Collector &c, void *p, size_t size) {
 		// → std::free に流れて整合する。
 		return std::realloc(p, size);
 	}
-	size_t old_size = oldH->size;
+	size_t old_size = static_cast<size_t>(oldH->size);
 	void *q = do_malloc(c, size);
 	if (!q) return nullptr; // 元の p は触らずそのまま (標準準拠)
 	std::memcpy(q, p, old_size < size ? old_size : size);
